Refactored MultiContest into its own file
